openvm fib bench

+++ openvm/fibonacci-guest/src/main.rs
@@ -6,11 +6,11 @@ use openvm::io::{read, reveal};
 openvm::entry!(main);
 
 pub fn main() {
-    let n: u64 = read();
-    let mut a: u64 = 0;
-    let mut b: u64 = 1;
+    let n: u32 = read();
+    let mut a: u128 = 0;
+    let mut b: u128 = 1;
     for _ in 0..n {
-        let c: u64 = a.wrapping_add(b);
+        let c: u128 = a.wrapping_add(b);
         a = b;
         b = c;
     }
